refactor rest models: move away from grpc crate
