+++ lib/api/src/grpc/mod.rs
@@ -1,5 +1,4 @@
 pub mod conversions;
-pub mod models;
 #[allow(clippy::all)]
 #[rustfmt::skip] // tonic uses `prettyplease` to format its output
 pub mod qdrant;
