+++ lib/api/src/rest/mod.rs
@@ -1,5 +1,6 @@
 pub mod conversions;
 pub mod schema;
 pub mod validate;
+pub mod models;
 
 pub use schema::*;
