+++ src/greeting.rs
@@ -3,7 +3,7 @@ use std::env;
 use std::io::{stdout, IsTerminal};
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
-use api::grpc::models::get_git_commit_id;
+use api::rest::models::get_git_commit_id;
 use colored::{Color, ColoredString, Colorize};
 
 use crate::settings::Settings;
